Add files via upload
EfficientNetV2 is an improved version of the EfficientNet model, which is known for its efficiency in terms of accuracy and computational resources. It incorporates new techniques such as the Swish activation function and the Squeeze-and-Excitation (SE) block to further enhance performance.

ResNet, on the other hand, is a popular deep learning model known for its use of residual connections to address the vanishing gradient problem. It has been widely used in various computer vision tasks and has proven to be effective in practice.